% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparty.R
\name{geom_edge_label_continuous}
\alias{geom_edge_label_continuous}
\title{generate geom of continuous splitvars}
\usage{
geom_edge_label_continuous(mapping = NULL, ...)
}
\arguments{
\item{mapping}{not recommended to change}
}
\description{
generate geom of continuous splitvars
}
