% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparty.R
\name{geom_edge_label_discrete}
\alias{geom_edge_label_discrete}
\title{generate geom of discrete splitvars}
\usage{
geom_edge_label_discrete(mapping = NULL, ...)
}
\arguments{
\item{mapping}{not recommended to change}
}
\description{
generate geom of discrete splitvars
}
