% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparty.R
\name{geom_node_info}
\alias{geom_node_info}
\title{Draw labels containing node's info}
\usage{
geom_node_info(mapping = NULL, x_nudge = 0, y_nudge = 0,
  ids = NULL, ...)
}
\arguments{
\item{mapping}{not recommended to change}

\item{x_nudge, y_nudge}{nudge label}

\item{ids}{choose which nodes to label by their ids}

\item{...}{additional arguments for \code{\link[=geom_label]{geom_label()}}}
}
\description{
Draw labels containing node's info
}
