% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparty.R
\name{geom_node_inner}
\alias{geom_node_inner}
\title{generate geom of inner nodes}
\usage{
geom_node_inner(mapping = NULL, ...)
}
\arguments{
\item{mapping}{not recommended to change}
}
\description{
generate geom of inner nodes
}
