% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparty.R
\name{geom_node_terminal_label}
\alias{geom_node_terminal_label}
\title{generate geom of terminal nodes}
\usage{
geom_node_terminal_label(mapping = NULL, ...)
}
\arguments{
\item{mapping}{not recommended to change}
}
\description{
generate geom of terminal nodes
}
