% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparty.R
\name{geom_edge}
\alias{geom_edge}
\title{Draw edges between children and parents}
\usage{
geom_edge(mapping = NULL, nudge_x = 0, nudge_y = 0, ids = NULL,
  ...)
}
\arguments{
\item{mapping}{not recommended to change}

\item{nudge_x, nudge_y}{nudge label}

\item{ids}{choose which edges to draw by their children's ids}

\item{...}{additional arguments for \code{\link[=geom_segment]{geom_segment()}}}
}
\description{
wrapper of geom_segment. Draws edges between child and parent nodes.
}
