% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparty.R
\name{geom_node_info}
\alias{geom_node_info}
\title{Draw labels containing node's info}
\usage{
geom_node_info(mapping = NULL, nudge_x = 0, nudge_y = 0,
  ids = NULL, label.padding = unit(0.5, "lines"), ...)
}
\arguments{
\item{mapping}{not recommended to change}

\item{nudge_x, nudge_y}{nudge label}

\item{ids}{choose which nodes to label by their ids}

\item{label.padding}{Amount of padding around label. Defaults to 0.5 lines.}

\item{...}{additional arguments for \code{\link[=geom_label]{geom_label()}}}
}
\description{
Draw labels containing node's info
}
