% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_nodeplot.R
\name{geom_nodeplot}
\alias{geom_nodeplot}
\title{generate geom of nodeplot}
\usage{
geom_nodeplot(plot_call = "ggplot", gglist = NULL, width = 0.1,
  height = 0.1, ids = NA, scales = "fixed", x_nudge = 0,
  y_nudge = 0)
}
\arguments{
\item{plot_call}{any function that generates a ggplot2 object.}

\item{gglist}{list of additional ggplot components. Data of nodes can be
mapped. Fitted values of modelparty objects can be mapped with "fitted_values".}

\item{width, height}{size of the nodeplot's viewport}

\item{ids}{which ids to plot. numeric or string "terminal"}

\item{scales}{see \code{\link[=facet_wrap]{facet_wrap()}}}

\item{x_nudge, y_nudge}{nudge nodeplot}
}
\description{
generate geom of nodeplot
}
