% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_nodeplot.R
\name{geom_nodeplot}
\alias{geom_nodeplot}
\title{generate geom of nodeplot}
\usage{
geom_nodeplot(mapping = NULL, data = NULL, gglist = NULL,
  plot_call = "ggplot", width = 0.1, height = 0.1, ids = NA,
  scales = "fixed", x_nudge = 0, y_nudge = 0, ...)
}
\arguments{
\item{mapping}{mapping of nodeplots. if NULL uses inherited data. fitted
values of modelparty objects can be mapped with "fitted_values".}

\item{gglist}{list of additional ggplot components}

\item{width, height}{size of the nodeplot's viewport}

\item{ids}{which ids to plot. numeric or string "terminal"}

\item{scales}{see \code{\link[=facet_wrap]{facet_wrap()}}}

\item{xnudge, ynudge}{nudge nodeplot}
}
\description{
generate geom of nodeplot
}
