% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_nodeplot.R
\name{geom_nodeplot}
\alias{geom_nodeplot}
\title{generate geom of nodeplot}
\usage{
geom_nodeplot(plot_call = "ggplot", gglist = NULL, width = 1,
  height = 1, ids = "terminal", scales = "fixed", x_nudge = 0,
  y_nudge = 0, shared_axes_labels = FALSE)
}
\arguments{
\item{plot_call}{any function that generates a ggplot2 object.}

\item{gglist}{list of additional ggplot components. Data of nodes can be
mapped. Fitted values of modelparty objects can be mapped with "fitted_values".}

\item{width}{expansion factor for viewport's width}

\item{height}{expansion factor for viewport's height}

\item{ids}{which ids to plot. numeric, "terminal", "inner" ar "all". defaults
to "terminal"}

\item{scales}{see \code{\link[=facet_wrap]{facet_wrap()}}}

\item{x_nudge, y_nudge}{nudge nodeplot}

\item{shared_axes_labels}{if TRUE only one pair of axes labels is plotted in
the terminal space. Only recommended if \code{ids}  "terminal" or "all".}
}
\description{
generate geom of nodeplot
}
\seealso{
\code{\link[=ggparty]{ggparty()}}
}
