% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparty.R
\name{ggparty}
\alias{ggparty}
\title{ggplot an object of the party class}
\usage{
ggparty(party, horizontal = FALSE)
}
\arguments{
\item{party}{partyobject to plot}

\item{horizontal}{horizontal plot?}
}
\description{
ggplot an object of the party class
}
