% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparty.R
\name{ggparty}
\alias{ggparty}
\title{ggplot an object of the party class}
\usage{
ggparty(party, horizontal = FALSE, terminal_space = 0.2,
  layout = NULL)
}
\arguments{
\item{party}{partyobject to plot}

\item{horizontal}{horizontal plot?}

\item{terminal_space}{proportion of the plot that should be reserved for
the terminal nodeplots}

\item{layout}{optional layout adjustment. Must be data.frame containing the
columns "id", "x" and "y".}
}
\description{
ggplot an object of the party class
}
\seealso{
\code{\link[=geom_edge]{geom_edge()}}, \code{\link[=geom_edge_label]{geom_edge_label()}}, \code{\link[=geom_node_splitvar]{geom_node_splitvar()}},
\code{\link[=geom_node_info]{geom_node_info()}}, \code{\link[=geom_nodeplot]{geom_nodeplot()}}
}
