% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_node_plot.R
\name{makeContent.nodeplotgrob}
\alias{makeContent.nodeplotgrob}
\title{apparantly needs to be exported}
\usage{
\method{makeContent}{nodeplotgrob}(x)
}
\arguments{
\item{x}{nodeplotgrob}
}
\description{
apparantly needs to be exported
}
